requirements.txt
# LLM clients
groq>=0.13.0

# Config
python-dotenv>=1.0.0

# Data engineering
pandas>=2.2.0
pyarrow>=15.0.0
requests>=2.32.0

# Validation
pydantic>=2.7.0

# AWS SDK (runtime for Lambda, also useful for local dev scripts)
boto3>=1.34.0

# v3.5 — Google Cloud Storage + Azure Blob connectors
google-cloud-storage>=2.10.0
azure-storage-blob>=12.19.0

# v3 persistence + auth + email
sqlalchemy>=2.0.30
alembic>=1.13.0
psycopg2-binary>=2.9.9       # Postgres driver (Supabase, Render Postgres, etc.)
python-jose[cryptography]>=3.3.0   # JWT (v3.1)
bcrypt>=4.0.0                # password hashing (passlib removed — buggy with bcrypt 4.x)
resend>=2.0.0                # transactional email (password reset)

# v3.7 — error tracking. Optional at runtime: if SENTRY_DSN isn't set we
# skip init, so this is safe to ship even on the free Render tier.
sentry-sdk[fastapi]>=1.40.0

# Dev / testing
pytest>=8.0.0

# Demo UI
streamlit>=1.36.0

# v2 backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
# streamlit-authenticator removed in v3 — replaced by API-backed auth.
# Kept as a comment to avoid `pip uninstall` confusion on existing venvs:
#   streamlit-authenticator>=0.3.3
